WIP: move connection config type readers into its own namespace
