+++ src/lib/y2network/config_reader/connection_config/sysconfig_handlers/wlan.rb
@@ -17,30 +17,36 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require "y2network/connection_config/ethernet"
+require "y2network/connection_config/wireless"
 
 module Y2Network
   module ConfigReader
     module ConnectionConfig
-      # This class is able to build a ConnectionConfig::Ethernet object given a
+      # This class is able to build a ConnectionConfig::Wireless object given a
       # SysconfigInterfaceFile object.
-      class SysconfigEthernet
+      class Wlan
         # @return [Y2Network::SysconfigInterfaceFile]
         attr_reader :file
 
-        # Constructor
-        #
-        # @param file [Y2Network::SysconfigInterfaceFile] File to get interface configuration from
         def initialize(file)
           @file = file
         end
 
-        # @return [Y2Network::ConnectionConfig::Ethernet]
+        # Returns an ethernet connection configuration
+        #
+        # @param name [String] Interface name
+        # @return [ConnectionConfig::Ethernet]
         def connection_config
-          Y2Network::ConnectionConfig::Ethernet.new.tap do |conn|
+          Y2Network::ConnectionConfig::Wireless.new.tap do |conn|
             conn.interface = file.name
-            conn.bootproto = file.fetch("BOOTPROTO")
+            conn.bootproto = file.fetch("BOOTPROTO").to_sym
             conn.ip_address = file.ip_address
+            conn.essid = file.fetch("WIRELESS_ESSID")
+            conn.mode = file.fetch("WIRELESS_MODE")
+            conn.auth_mode = file.fetch("WIRELESS_AUTH_MODE")
+            conn.wpa_psk = file.fetch("WIRELESS_WPA_PSK")
+            # TODO: conn.nwid
+            conn.wpa_psk = file.fetch("WIRELESS_KEY_LENGTH")
           end
         end
       end
